Fixing Categories and Tasks Linking

+++ src/com/android/casorio/categories/IOnCategorySelectedListener.java
@@ -3,6 +3,6 @@ package com.android.casorio.categories;
 public interface IOnCategorySelectedListener {
 	
 	
-	public void onCategorySelected(int position);
+	public void onCategorySelected(long position);
 
 }
